pyproject.toml
[project]
name = "maths-bot"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "python-telegram-bot>=22.5",
]
